requirements.txt
pyspark
findspark
kafka-python
notebook